+++ src/piecewise_exponential_rate_function.h
@@ -37,7 +37,7 @@ class PiecewiseExponentialRateFunction
     Matrix<T> tjj_all_above(const int, const MatrixXq&, const MatrixXq&, const MatrixXq&, const MatrixXq&) const;
     void tjj_double_integral_above(const int, long, std::vector<Matrix<T> > &) const;
     void tjj_double_integral_below(const int, const int, Matrix<T>&) const;
-    double random_time(const double, const double, const int) const;
+    double random_time(const double, const double, const long long) const;
     T random_time(const double, const T&, const T&, std::mt19937&) const;
 
     friend class ConditionedSFS<T>;
